revert: drop the drizzle migration system from studio
