--- apps/studio/server/drizzle/meta/_journal.json
@@ -1,13 +1,0 @@
-{
-  "version": "7",
-  "dialect": "postgresql",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "7",
-      "when": 1786568320900,
-      "tag": "0000_nervous_sister_grimm",
-      "breakpoints": true
-    }
-  ]
-}
