+++ apps/studio/server/src/auth/better-auth.ts
@@ -1,10 +1,7 @@
 import { betterAuth } from 'better-auth';
-import { drizzleAdapter } from 'better-auth/adapters/drizzle';
 import { magicLink } from 'better-auth/plugins';
-import { drizzle } from 'drizzle-orm/node-postgres';
 import type pg from 'pg';
 
-import * as schema from '../db/schema.ts';
 import type { AuthEnv } from '../env.ts';
 import type { MagicLinkMailer } from './email.ts';
 import type { AuthService } from './service.ts';
@@ -23,13 +20,7 @@ export function createBetterAuthInstance(
     baseURL: env.baseUrl,
     basePath: '/api/auth',
     secret: env.secret,
-    // The drizzle adapter over src/db/schema.ts — the same schema the boot
-    // migrations create, so the ORM shape and the database shape cannot
-    // drift silently.
-    database: drizzleAdapter(drizzle(pool, { schema }), {
-      provider: 'pg',
-      schema,
-    }),
+    database: pool,
     // better-auth's own CSRF for /api/auth/*; the rest of the cookie plane
     // is covered by src/auth/csrf.ts (#1248).
     trustedOrigins: [env.baseUrl],
