+++ apps/studio/server/src/db/schema.ts
@@ -1,110 +1,37 @@
-import {
-  bigint,
-  boolean,
-  index,
-  integer,
-  pgTable,
-  text,
-  timestamp,
-} from 'drizzle-orm/pg-core';
+import type pg from 'pg';
 
 // better-auth's tables (users, cookie sessions, provider accounts,
 // verification tokens) plus the Postgres-backed rate-limit counters (#1246:
 // durable security counters live in Postgres, never memory or Redis).
 //
-// This is the Drizzle source of truth: better-auth reads it through the
-// drizzle adapter (src/auth/better-auth.ts) and drizzle-kit generates the
-// SQL migrations in drizzle/ from it (`pnpm db:generate`), which boot
-// applies (src/index.ts). Column names are deliberately camelCase (quoted
-// identifiers) — better-auth's default shape, kept stable across the move
-// from the pre-migration snapshot SQL.
-//
-// On a better-auth version bump: re-run `npx -y @better-auth/cli@latest
-// generate --config scripts/auth-cli-config.ts`, diff the emitted schema
-// against this file, fold in any changes, then `pnpm db:generate` and
-// review the new migration before committing it.
+// Snapshotted from `npx -y @better-auth/cli@latest generate --config
+// scripts/auth-cli-config.ts` (CLI 1.4.22 at snapshot time — the CLI
+// versions independently of better-auth itself) and made idempotent, the
+// same shape as
+// packages/studio-sync's SCHEMA_SQL. There is deliberately no migration
+// system yet: pre-release, a schema change means wiping the database and
+// letting boot recreate it. Real migrations (Drizzle, per #1246's direction)
+// must land before a release carries data worth keeping. On a better-auth
+// version bump: re-run generate against an empty database, diff against
+// this, and update.
+const AUTH_SCHEMA_SQL = `
+create table if not exists "user" ("id" text not null primary key, "name" text not null, "email" text not null unique, "emailVerified" boolean not null, "image" text, "createdAt" timestamptz default CURRENT_TIMESTAMP not null, "updatedAt" timestamptz default CURRENT_TIMESTAMP not null);
 
-export const user = pgTable('user', {
-  id: text('id').primaryKey(),
-  name: text('name').notNull(),
-  email: text('email').notNull().unique(),
-  emailVerified: boolean('emailVerified').notNull(),
-  image: text('image'),
-  createdAt: timestamp('createdAt', { withTimezone: true })
-    .defaultNow()
-    .notNull(),
-  updatedAt: timestamp('updatedAt', { withTimezone: true })
-    .defaultNow()
-    .notNull(),
-});
+create table if not exists "session" ("id" text not null primary key, "expiresAt" timestamptz not null, "token" text not null unique, "createdAt" timestamptz default CURRENT_TIMESTAMP not null, "updatedAt" timestamptz not null, "ipAddress" text, "userAgent" text, "userId" text not null references "user" ("id") on delete cascade);
 
-export const session = pgTable(
-  'session',
-  {
-    id: text('id').primaryKey(),
-    expiresAt: timestamp('expiresAt', { withTimezone: true }).notNull(),
-    token: text('token').notNull().unique(),
-    createdAt: timestamp('createdAt', { withTimezone: true })
-      .defaultNow()
-      .notNull(),
-    updatedAt: timestamp('updatedAt', { withTimezone: true }).notNull(),
-    ipAddress: text('ipAddress'),
-    userAgent: text('userAgent'),
-    userId: text('userId')
-      .notNull()
-      .references(() => user.id, { onDelete: 'cascade' }),
-  },
-  (t) => [index('session_userId_idx').on(t.userId)],
-);
+create table if not exists "account" ("id" text not null primary key, "accountId" text not null, "providerId" text not null, "userId" text not null references "user" ("id") on delete cascade, "accessToken" text, "refreshToken" text, "idToken" text, "accessTokenExpiresAt" timestamptz, "refreshTokenExpiresAt" timestamptz, "scope" text, "password" text, "createdAt" timestamptz default CURRENT_TIMESTAMP not null, "updatedAt" timestamptz not null);
 
-export const account = pgTable(
-  'account',
-  {
-    id: text('id').primaryKey(),
-    accountId: text('accountId').notNull(),
-    providerId: text('providerId').notNull(),
-    userId: text('userId')
-      .notNull()
-      .references(() => user.id, { onDelete: 'cascade' }),
-    accessToken: text('accessToken'),
-    refreshToken: text('refreshToken'),
-    idToken: text('idToken'),
-    accessTokenExpiresAt: timestamp('accessTokenExpiresAt', {
-      withTimezone: true,
-    }),
-    refreshTokenExpiresAt: timestamp('refreshTokenExpiresAt', {
-      withTimezone: true,
-    }),
-    scope: text('scope'),
-    password: text('password'),
-    createdAt: timestamp('createdAt', { withTimezone: true })
-      .defaultNow()
-      .notNull(),
-    updatedAt: timestamp('updatedAt', { withTimezone: true }).notNull(),
-  },
-  (t) => [index('account_userId_idx').on(t.userId)],
-);
+create table if not exists "verification" ("id" text not null primary key, "identifier" text not null, "value" text not null, "expiresAt" timestamptz not null, "createdAt" timestamptz default CURRENT_TIMESTAMP not null, "updatedAt" timestamptz default CURRENT_TIMESTAMP not null);
 
-export const verification = pgTable(
-  'verification',
-  {
-    id: text('id').primaryKey(),
-    identifier: text('identifier').notNull(),
-    value: text('value').notNull(),
-    expiresAt: timestamp('expiresAt', { withTimezone: true }).notNull(),
-    createdAt: timestamp('createdAt', { withTimezone: true })
-      .defaultNow()
-      .notNull(),
-    updatedAt: timestamp('updatedAt', { withTimezone: true })
-      .defaultNow()
-      .notNull(),
-  },
-  (t) => [index('verification_identifier_idx').on(t.identifier)],
-);
+create table if not exists "rateLimit" ("id" text not null primary key, "key" text not null unique, "count" integer not null, "lastRequest" bigint not null);
 
-export const rateLimit = pgTable('rateLimit', {
-  id: text('id').primaryKey(),
-  key: text('key').notNull().unique(),
-  count: integer('count').notNull(),
-  lastRequest: bigint('lastRequest', { mode: 'number' }).notNull(),
-});
+create index if not exists "session_userId_idx" on "session" ("userId");
+
+create index if not exists "account_userId_idx" on "account" ("userId");
+
+create index if not exists "verification_identifier_idx" on "verification" ("identifier");
+`;
+
+export async function applyAuthSchema(pool: pg.Pool): Promise<void> {
+  await pool.query(AUTH_SCHEMA_SQL);
+}
