+++ apps/studio/server/tsconfig.json
@@ -5,11 +5,6 @@
     "strict": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": [
-    "src/**/*.ts",
-    "drizzle.config.ts",
-    "vite.config.ts",
-    "vitest.config.ts"
-  ],
+  "include": ["src/**/*.ts", "vite.config.ts", "vitest.config.ts"],
   "exclude": ["node_modules", "dist"]
 }
